Backend/backendrequirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
textblob==0.17.1
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
openai>=1.0.0
ta>=0.10.2
alpha-vantage>=2.3.1
curl-cffi>=0.5.0